backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-q2==1.7.4
django-cors-headers==4.3.1
gunicorn==22.0.0
python-dotenv==1.0.1
